ai_scheduler/requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
